DOC: Rewrite software release instructions.
Most of this very good content from @tacaswell was captured in a more
detailed form in the cookiecutter docs. The few details that the
cookiecutter will be added in this new PR:

https://github.com/NSLS-II/scientific-python-cookiecutter/pull/41

Some details (mention of "binstar") were very out of date.

The new page links out to the cookiecutter and just adds the NSLS-II
specific stuff.